+++ src/ground_facilities/ground_facility_accesses_and_gaps.jl
@@ -65,7 +65,7 @@ Those geodetic information are transformed to an ECEF vector using the function
 - `unit::Symbol`: Select the unit in which the duration will be computed. The possible
     values are:
     - `:s` for seconds (**Default**);
-    - `:m` for minutes; or
+    - `:min` for minutes; or
     - `:h` for hours.
 
 # Returns
@@ -80,7 +80,7 @@ Those geodetic information are transformed to an ECEF vector using the function
 
 ## Throws
 
-- `ArgumentError`: If `unit` is not `:s`, `:m`, or `:h`.
+- `ArgumentError`: If `unit` is not `:s`, `:min`, or `:h`.
 
 ## Examples
 
@@ -109,7 +109,7 @@ julia> ground_facility_accesses(orbp, (0, 0, 0))
    1 │ 2024-01-01T10:20:03.136  2024-01-01T10:30:02.971   599.835
    2 │ 2024-01-01T22:49:55.910  2024-01-01T22:59:23.470   567.56
 
-julia> ground_facility_accesses(orbp, (0, 0, 0); unit = :m)
+julia> ground_facility_accesses(orbp, (0, 0, 0); unit = :min)
 2×3 DataFrame
  Row │ access_beginning         access_end               duration 
      │ DateTime                 DateTime                 Float64  
@@ -274,7 +274,7 @@ lasts for `duration` [s].
 
 ## Throws
 
-- `ArgumentError`: If `unit` is not `:s`, `:m`, or `:h`.
+- `ArgumentError`: If `unit` is not `:s`, `:min`, or `:h`.
 
 ## Examples
 
@@ -304,7 +304,7 @@ julia> ground_facility_gaps(orbp, (0, 0, 0))
    2 │ 2024-01-01T10:30:02.971  2024-01-01T22:49:55.910  44392.9
    3 │ 2024-01-01T22:59:23.470  2024-01-02T00:00:00       3636.53
 
-julia> ground_facility_gaps(orbp, (0, 0, 0); unit = :m)
+julia> ground_facility_gaps(orbp, (0, 0, 0); unit = :min)
 3×3 DataFrame
  Row │ gap_beginning            gap_end                  duration 
      │ DateTime                 DateTime                 Float64